requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
psycopg2-binary
pandas
openpyxl
Werkzeug
